feat(bs): add machine bootstrap playbook
Add a playbook to bootstrap the Ansible setup with machines in the
bootstrap group. The idea is for machines that haven't yet been set up
to be placed under the list of hosts in the 'bootstrap' group.

The last task of this playbook is to remove the host from the bootstrap
group by invoking yq. The assumption is that the inventory file is YAML
formatted, and that yq is installed on the host machine.