feat(bs): allow sudo group to sudo w/o passwd
Allow the sudo group to become root without need for a password, since
it would be a hassle to manually set passwords for each user as created
by the users playbook.

Note that there's no task on this repository that makes any users part
of the sudo group. That is by design: allowing a user into the sudo
group is a rarely performed task, so it will be carried out through
ad-hoc commands.