refactor: remove -playbook suffix from playbooks
Considering the name of this repository is 'playbooks' (at least on
GitHub) and executing them involves calling ansible-playbook, it's a
little obvious these YAML files are Ansible playbooks... right?

This could be convention and I might be making a mistake, but you can't
say it doesn't look nicer than having '-playbook' at the end of every
playbook filename.