feat: enable pipelining to reuse SSH connection
Enable SSH pipelining on Ansible to try and reuse a single SSH
connection as allowed by the settings on modules. This way hopefully
plays take less time to complete.

This is being done to ensure firewalld won't get in the way of the task
that ensures the configured SSH port will be closed.